Modifiable Propulsion!
Victory!

I re-coded the Propulsion.m Class definition, to skip the
propulsionTypes.dat to get the numbers.  This will allow multiple
instances of the simulation to run in parallel, with different
propulsion numbers.  The values are now all intrinsic to the object.
Defaults in the constructor function in Propulsion.m